+++ lib/fun/comics.js
@@ -19,7 +19,7 @@ const help_text = `Get random web comic strips
 **\`!comic cah\`** or **\`!comic cyanide and happiness\`**
     Returns a Cyanide and Happimess comic
 **\`!comic css\`** or **\`!comic chainsawsuit\`**
-    Returns a Chainsawsuit comic
+    Returns a chainsawsuit comic
 **\`!comic dhd\`** or **\`!comic dog house diaries\`**
     Returns a Dog House Diaries comic
 **\`!comic smbc\`** or ** \`!comic saturday morning breakfast cereal\`**
