+++ tests/fun/translate.js
