+++ tests/index.js
@@ -1,5 +1,6 @@
 import glob from 'glob';
 import R from 'ramda';
+
 import '../lib/config/init';
 import '../lib/config/express';
 import '../lib/config/phantom';
